Re-implemented metronome
+ Added a new metronome using a shader and ties key press to the beat
   note : Currently very inconsistent and frustrating, will need to be revisited

TO DO:
-> Add a graphical visualization helper of the different keys around the player

+++ Assets/Scripts/Music/MusicUtils.cs
@@ -9,8 +9,8 @@ namespace MusicNamespace
     public struct MusicUtils
     {
 
-        //should not be here
-        static public int BPM = 60;
+        //should not be here?
+        static public float BPM = 60f;
 
         private readonly static int[] intervals = { 2, 2, 1, 2, 2, 2, 1, 2, 2, 1, 2, 2, 2, 1}; // Whole and half steps for Ionian mode over two octave
